pyproject.toml
[project]
name = "ai-travel-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.7",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.5",
    "langchain-groq>=1.0.1",
    "langchain-ollama>=1.0.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
